requirements.txt
tqdm
joblib
ttkbootstrap
numpy
pandas
pillow
opencv-python >= 4.5.1
scikit-image >= 0.18.1
torch >= 1.8
torchvision
